+++ lib/libsesame3bt-core/debug.h
@@ -3,7 +3,6 @@
 #if LIBSESAME3BTCORE_DEBUG
 #ifdef ARDUINO
 #include <Arduino.h>
-#endif
 #define DEBUG_PRINTLN(...)      \
 	do {                          \
 		Serial.printf(__VA_ARGS__); \
@@ -17,7 +16,22 @@
 	do {                          \
 		Serial.printf(__VA_ARGS__); \
 	} while (false)
-
+#else
+#include <stdio.h>
+#define DEBUG_PRINTLN(...) \
+	do {                     \
+		printf(__VA_ARGS__);   \
+		putchar('\n');         \
+	} while (false)
+#define DEBUG_PRINT(...) \
+	do {                   \
+		printf(__VA_ARGS__); \
+	} while (false)
+#define DEBUG_PRINTF(...) \
+	do {                    \
+		printf(__VA_ARGS__);  \
+	} while (false)
+#endif
 #else
 
 #define DEBUG_PRINTLN(...) \